+++ test/features/agents/state/change_set_providers_test.dart
@@ -14,6 +14,7 @@ import 'package:lotti/features/tasks/repository/checklist_repository.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/providers/service_providers.dart' show journalDbProvider;
+import 'package:lotti/services/time_service.dart';
 import 'package:mocktail/mocktail.dart';
 
 import '../../../helpers/fallbacks.dart';
@@ -321,7 +322,9 @@ void main() {
       final mockLabelsRepository = MockLabelsRepository();
       final mockPersistenceLogic = MockPersistenceLogic();
 
-      getIt.registerSingleton<PersistenceLogic>(mockPersistenceLogic);
+      getIt
+        ..registerSingleton<PersistenceLogic>(mockPersistenceLogic)
+        ..registerSingleton<TimeService>(TimeService());
       addTearDown(() async => getIt.reset());
 
       final container = ProviderContainer(